Refactor tsbuild watch tests (#44258)
