+++ src/testRunner/tsconfig.json
@@ -135,8 +135,13 @@
         "unittests/tsbuild/resolveJsonModule.ts",
         "unittests/tsbuild/sample.ts",
         "unittests/tsbuild/transitiveReferences.ts",
-        "unittests/tsbuild/watchEnvironment.ts",
-        "unittests/tsbuild/watchMode.ts",
+        "unittests/tsbuildWatch/configFileErrors.ts",
+        "unittests/tsbuildWatch/demo.ts",
+        "unittests/tsbuildWatch/moduleResolution.ts",
+        "unittests/tsbuildWatch/noEmitOnError.ts",
+        "unittests/tsbuildWatch/programUpdates.ts",
+        "unittests/tsbuildWatch/reexport.ts",
+        "unittests/tsbuildWatch/watchEnvironment.ts",
         "unittests/tsc/composite.ts",
         "unittests/tsc/declarationEmit.ts",
         "unittests/tsc/incremental.ts",
